+++ oscqam/__init__.py
@@ -1,5 +1,5 @@
 from distutils.version import StrictVersion
 
-__version__ = '0.4.2'
+__version__ = '0.5.0'
 
 strict_version = StrictVersion(__version__)