+++ oscqam/actions.py
@@ -37,7 +37,7 @@ class NoQamReviewsError(UninferableError):
         message += (" The following groups were already assigned/finished: "
                     "{msg}".format(
                         msg=", ".join(["{r.reviewer}".format(
-                            r=review
+                            r = review
                         ) for review in accept_reviews])
                     )) if accept_reviews else ""
         super(NoQamReviewsError, self).__init__(message)
@@ -94,6 +94,29 @@ class TestResultMismatchError(ReportedError):
     def __init__(self, expected, log_path):
         super(TestResultMismatchError, self).__init__(
             self._msg.format(expected, log_path)
+        )
+
+
+class ReportNotYetGeneratedError(ReportedError):
+    _msg = ("The report for request '{0}' is not generated yet. "
+            "To prevent bugs in the template parser, assigning "
+            "is not yet possible.")
+
+    def __init__(self, request):
+        super(ReportNotYetGeneratedError, self).__init__(
+            self._msg.format(str(request))
+        )
+
+
+class OneGroupAssignedError(ReportedError):
+    _msg = ("User {user} is already assigned for group {group}. "
+            "Assigning for multiple groups at once is currently not allowed "
+            "to prevent inconsistent states in the build service.")
+
+    def __init__(self, assignment):
+        super(OneGroupAssignedError, self).__init__(
+            self._msg.format(user = str(assignment.user),
+                             group = str(assignment.group))
         )
 
 
@@ -166,8 +189,8 @@ class ListAction(OscAction):
                     logger.debug("Missing key: %s", key)
             return data
 
-    def __init__(self, remote, user, only_review=False,
-                 template_factory=Template):
+    def __init__(self, remote, user, only_review = False,
+                 template_factory = Template):
         super(ListAction, self).__init__(remote, user)
         self.only_review = only_review
         self.template_factory = template_factory
@@ -236,10 +259,43 @@ class AssignAction(OscAction):
     AUTO_INFER_MSG = "Found a possible group: {group}."
     MULTIPLE_GROUPS_MSG = "User could review more than one group: {groups}"
 
-    def __init__(self, remote, user, request_id, group = None):
+    def __init__(self, remote, user, request_id, group = None,
+                 template_factory = Template):
         super(AssignAction, self).__init__(remote, user)
         self.request = Request.by_id(self.remote, request_id)
         self.group = Group.for_name(remote, group) if group else None
+        self.template_factory = template_factory
+
+    def template_exists(self):
+        """Check that the template associated with the request exists.
+
+        If the template is not yet generated, assigning a user can lead
+        to the template-generator no longer finding the request and
+        never generating the template.
+        """
+        try:
+            self.request.get_template(self.template_factory)
+        except TemplateNotFoundError:
+            raise ReportNotYetGeneratedError(self.request)
+
+    def first_group_assigned(self):
+        """Prevent a user from assigned more than 1 group at a time.
+
+        The buildservice creates only one user review, even if the user
+        assigns to 2 groups.
+        This user-review will be closed as soon as the user accepts the
+        first review, leading to inconsistent state in the buildservice.
+
+        Not allowing a user to assign for > 1 group at a time at least
+        prevents this.
+        """
+        for assignment in self.request.assigned_roles:
+            if assignment.user == self.user:
+                raise OneGroupAssignedError(assignment)
+
+    def validate(self):
+        self.template_exists()
+        self.first_group_assigned()
 
     def action(self):
         if self.group:
@@ -276,6 +332,7 @@ class AssignAction(OscAction):
         return group
 
     def assign(self, group):
+        self.validate()
         msg = AssignAction.ASSIGN_USER_MSG.format(
             user = self.user, group = group, request = self.request
         )
@@ -284,8 +341,8 @@ class AssignAction(OscAction):
             prefix = PREFIX, user = self.user, group = group
         )
         self.request.review_assign(reviewer = self.user,
-                                   group = group,
-                                   comment = comment)
+                                    group = group,
+                                    comment = comment)
 
 
 class UnassignAction(OscAction):