+++ tests/test_actions.py
@@ -24,6 +24,7 @@ class ActionTests(unittest.TestCase):
         self.non_open = '23456'
         self.sle_open = '34567'
         self.non_qam = '45678'
+        self.one_assigned = '56789'
         self.assigned = '52542'
         self.single_assign_single_open = 'oneassignoneopen'
         self.two_assigned = 'twoassigned'
@@ -42,7 +43,8 @@ class ActionTests(unittest.TestCase):
 
     def test_infer_groups_match(self):
         assign_action = actions.AssignAction(self.mock_remote, self.user_id,
-                                             self.sle_open)
+                                             self.sle_open,
+                                             template_factory = lambda r: True)
         assign_action()
         self.assertEqual(len(self.mock_remote.post_calls), 1)
 
@@ -65,18 +67,21 @@ class ActionTests(unittest.TestCase):
 
     def test_assign_non_matching_groups(self):
         assign = actions.AssignAction(self.mock_remote, self.user_id,
-                                      self.single_assign_single_open)
+                                      self.single_assign_single_open,
+                                      template_factory=lambda r: True)
         self.assertRaises(actions.NonMatchingGroupsError, assign)
 
     def test_assign_multiple_groups(self):
         assign = actions.AssignAction(self.mock_remote, self.user_id,
-                                      self.multi_available_assign)
+                                      self.multi_available_assign,
+                                      template_factory=lambda r: True)
         self.assertRaises(actions.UninferableError, assign)
 
     def test_assign_multiple_groups_explicit(self):
         assign = actions.AssignAction(self.mock_remote, self.user_id,
                                       self.multi_available_assign,
-                                      group='qam-test')
+                                      group='qam-test',
+                                      template_factory=lambda r: True)
         assign()
 
     def test_unassign_no_group(self):
@@ -97,3 +102,19 @@ class ActionTests(unittest.TestCase):
                                       self.cloud_open)
         action._template = template
         self.assertRaises(actions.TestResultMismatchError, action)
+
+    def test_assign_no_report(self):
+        def raiser(request):
+            raise models.TemplateNotFoundError("")
+        assign = actions.AssignAction(self.mock_remote, self.user_id,
+                                      self.multi_available_assign,
+                                      group = 'qam-test',
+                                      template_factory = raiser)
+        self.assertRaises(actions.ReportNotYetGeneratedError, assign)
+
+    def test_assign_only_one_group(self):
+        assign = actions.AssignAction(self.mock_remote, self.user_id,
+                                      self.one_assigned, group='qam-test',
+                                      template_factory=lambda r: True)
+        self.assertRaises(actions.OneGroupAssignedError, assign)
+